+++ submissions/Daniel-Badura/Readme.md
@@ -1,0 +1,42 @@
+# Spotify Data Logger
+
+## Overview
+
+The Spotify Data Logger is a Python script designed to read data from a CSV file containing Spotify music data and perform various analysis tasks. This project offers flexibility for adaptation to other backend programming languages. You can use this script as a starting point to analyze Spotify data for your specific needs.
+
+### Challenges
+
+1. Count the total number of songs in the CSV file.
+
+2. Identify the number of songs in the key of E.
+<!--
+3. Count the occurrences of values in a specified column (e.g., artist names) and determine the most common value.
+   -->
+
+## Prerequisites
+
+Before using this project, ensure you have the following prerequisites:
+
+- Python (any version)
+- A CSV file containing Spotify data (e.g., "spotify-2023.csv") in the same directory as the script, or you can specify the file path in the script.
+- Basic knowledge of Python and CSV file handling.
+
+## Getting Started
+
+Follow these steps to get started with the Spotify Data Logger:
+
+1. Clone or Download: Clone or download this repository to your local machine.
+
+2. **Prerequisites**: Ensure you have Python installed on your machine.
+
+3. **Running the Python Script**: The provided Python script (`main.py`) can perform Challenge 1 and Challenge 2. To run the script:
+
+   ```bash
+   python main.py
+   ```
+
+   It will count the total number of songs in the CSV file and the number of songs in the key of E, as specified in Challenge 1 and Challenge 2.
+
+## License
+
+This project is licensed under the MIT License. You are free to modify and distribute the adapted code as needed for your own projects.
