+++ submissions/Daniel-Badura/main.py
@@ -1,27 +1,11 @@
-import csv
+from modules.count_songs import count_songs
+from modules.count_songs_key import count_songs_key
 
 csv_path = "spotify-2023.csv"
 
-
-def count_songs(path):
-    try:
-        with open(path, "r", newline="") as csv_file:
-            csv_reader = csv.reader(csv_file)
-            next(csv_reader)
-            song_counter = 0
-
-            for row in csv_reader:
-                song_counter += 1
-
-            return song_counter
-
-    except FileNotFoundError:
-        return -404
-    except Exception:
-        return -400
-
-
 if __name__ == "__main__":
+    
+    # Challenge One
     song_count = count_songs(csv_path)
 
     if song_count == -404:
@@ -30,3 +14,10 @@ if __name__ == "__main__":
         print("Something went wrong.")
     else:
         print(f"Number of songs : {song_count}")
+    
+    # Challenge Two
+    key_column = "key" # You can modify the column
+    key_value = "E"    # and the value to be counted
+
+    songs_in_key = count_songs_key(csv_path, key_column, key_value)
+    print(f"Number of songs in the key of E: {songs_in_key}")
